A class for representing the cBioPortal API protocol

The `cBioPortal` class is a representation of the cBioPortal API protocol that directly inherits from the `Service` class in the `AnVIL` package. For more information, see the 'AnVIL' package.

Details

This class takes the static API as provided at https://www.cbioportal.org/api/api-docs and creates an R object with the help from underlying infrastructure (i.e., 'rapiclient' and 'AnVIL') to give the user a unified representation of the API specification provided by the cBioPortal group. Users are not expected to interact with this class other than to use it as input to the functionality provided by the rest of the package.

See also

Examples

cBioPortal()

#> service: cBioPortal
#> tags(); use cbioportal$<tab completion>:
#> # A tibble: 61 x 3
#>    tag           operation                  summary                             
#>    <chr>         <chr>                      <chr>                               
#>  1 Cancer Types  getAllCancerTypesUsingGET  Get all cancer types                
#>  2 Cancer Types  getCancerTypeUsingGET      Get a cancer type                   
#>  3 Clinical Att… fetchClinicalAttributesUs… Fetch clinical attributes           
#>  4 Clinical Att… getAllClinicalAttributesI… Get all clinical attributes in the …
#>  5 Clinical Att… getAllClinicalAttributesU… Get all clinical attributes         
#>  6 Clinical Att… getClinicalAttributeInStu… Get specified clinical attribute    
#>  7 Clinical Data fetchAllClinicalDataInStu… Fetch clinical data by patient IDs …
#>  8 Clinical Data fetchClinicalDataUsingPOST Fetch clinical data by patient IDs …
#>  9 Clinical Data getAllClinicalDataInStudy… Get all clinical data in a study    
#> 10 Clinical Data getAllClinicalDataOfPatie… Get all clinical data of a patient …
#> # … with 51 more rows
#> tag values:
#>   Cancer Types, Clinical Attributes, Clinical Data, Copy Number
#>   Segments, Discrete Copy Number Alterations, Gene Panels, Generic
#>   Assays, Genes, Molecular Data, Molecular Profiles, Mutations,
#>   Patients, Sample Lists, Samples, Structural Variants, Studies,
#>   Treatments
#> schemas():
#>   AlleleSpecificCopyNumber, AndedPatientTreatmentFilters,
#>   AndedSampleTreatmentFilters, CancerStudy, CancerStudyTags
#>   # ... with 56 more elements