Download data from the cBioPortal API

Obtain a MultiAssayExperiment object for a particular gene panel, studyId, molecularProfileIds, and sampleListIds combination. Default molecularProfileIds and sampleListIds are set to NULL for including all data. This option is best for users who wish to obtain a section of the study data that pertains to a specific molecular profile and gene panel combination. For users looking to download the entire study data as provided by the https://cbioportal.org/datasets, refer to cBioDataPack.

cBioPortalData(
  api,
  studyId = NA_character_,
  genePanelId = NA_character_,
  genes = NA_character_,
  molecularProfileIds = NULL,
  sampleListId = NULL,
  sampleIds = NULL,
  by = c("entrezGeneId", "hugoGeneSymbol"),
  check_build = TRUE
)

Arguments

api

An API object of class `cBioPortal` from the `cBioPortal` function

studyId

character(1) Indicates the "studyId" as taken from `getStudies`

genePanelId

character(1) Identifies the gene panel, as obtained from the `genePanels` function

genes

character() Either Entrez gene identifiers or Hugo gene symbols. When included, the 'by' argument indicates the type of identifier provided and 'genePanelId' is ignored. Preference is given to Entrez IDs due to faster query responses.

molecularProfileIds

character() A vector of molecular profile IDs

sampleListId

character(1) A sample list identifier as obtained from `sampleLists()``

sampleIds

character() Sample identifiers

by

character(1) Either 'entrezGeneId' or 'hugoGeneSymbol' for row metadata (default: 'entrezGeneId')

check_build

logical(1L) Whether to check the build status of the studyId using an internal dataset. This argument should be set to FALSE if using alternative hostnames, e.g., 'pedcbioportal.kidsfirstdrc.org'

Value

A MultiAssayExperiment object

Details

We are able to succesfully represent 98 percent of the study identifiers as MultiAssayExperiment objects as obtained via cBioPortalData with the IMPACT341 genePanelId as the example gene panel. Datasets that currently fail to import can be seen in the getStudies(..., buildReport = TRUE) dataset under the "api_build" column. Note that changes to the cBioPortal API may affect this rate at any time. If you encounter any issues, please open a GitHub issue at the https://github.com/waldronlab/cBioPortalData/issues/ page with a fully reproducible example.

See also

Examples

cbio <- cBioPortal()

samps <- samplesInSampleLists(cbio, "acc_tcga_rppa")[[1]]

getGenePanelMolecular(
    cbio, molecularProfileIds = c("acc_tcga_rppa", "acc_tcga_linear_CNA"),
    samps
)
#> # A tibble: 92 × 7
#>    uniqueSampleKey               uniqu…¹ molec…² sampl…³ patie…⁴ studyId profi…⁵
#>    <chr>                         <chr>   <chr>   <chr>   <chr>   <chr>   <lgl>  
#>  1 VENHQS1PUi1BNUoyLTAxOmFjY190… VENHQS… acc_tc… TCGA-O… TCGA-O… acc_tc… TRUE   
#>  2 VENHQS1PUi1BNUozLTAxOmFjY190… VENHQS… acc_tc… TCGA-O… TCGA-O… acc_tc… TRUE   
#>  3 VENHQS1PUi1BNUo2LTAxOmFjY190… VENHQS… acc_tc… TCGA-O… TCGA-O… acc_tc… TRUE   
#>  4 VENHQS1PUi1BNUo3LTAxOmFjY190… VENHQS… acc_tc… TCGA-O… TCGA-O… acc_tc… TRUE   
#>  5 VENHQS1PUi1BNUo4LTAxOmFjY190… VENHQS… acc_tc… TCGA-O… TCGA-O… acc_tc… TRUE   
#>  6 VENHQS1PUi1BNUo5LTAxOmFjY190… VENHQS… acc_tc… TCGA-O… TCGA-O… acc_tc… TRUE   
#>  7 VENHQS1PUi1BNUpBLTAxOmFjY190… VENHQS… acc_tc… TCGA-O… TCGA-O… acc_tc… TRUE   
#>  8 VENHQS1PUi1BNUpQLTAxOmFjY190… VENHQS… acc_tc… TCGA-O… TCGA-O… acc_tc… TRUE   
#>  9 VENHQS1PUi1BNUpSLTAxOmFjY190… VENHQS… acc_tc… TCGA-O… TCGA-O… acc_tc… TRUE   
#> 10 VENHQS1PUi1BNUpTLTAxOmFjY190… VENHQS… acc_tc… TCGA-O… TCGA-O… acc_tc… TRUE   
#> # … with 82 more rows, and abbreviated variable names ¹​uniquePatientKey,
#> #   ²​molecularProfileId, ³​sampleId, ⁴​patientId, ⁵​profiled

acc_tcga <- cBioPortalData(
    cbio, by = "hugoGeneSymbol",
    studyId = "acc_tcga",
    genePanelId = "AmpliSeq",
    molecularProfileIds =
        c("acc_tcga_rppa", "acc_tcga_linear_CNA", "acc_tcga_mutations")
)
#> Error in eval(args[["api"]]): object 'cbio' not found