The R interface to the cBioPortal API Data Service

This section of the documentation lists the functions that allow users to access the cBioPortal API. The main representation of the API can be obtained from the `cBioPortal` function. The supporting functions listed here give access to specific parts of the API and allow the user to explore the API with individual calls. Many of the functions here are listed for documentation purposes and are recommended for advanced usage only. Users should only need to use the `cBioPortalData` main function to obtain data.

cBioPortal(
  hostname = "www.cbioportal.org",
  protocol = "https",
  api. = "/api/v2/api-docs",
  token = character()
)

getStudies(api, buildReport = FALSE)

clinicalData(api, studyId = NA_character_)

molecularProfiles(
  api,
  studyId = NA_character_,
  projection = c("SUMMARY", "ID", "DETAILED", "META")
)

fetchData(
  api,
  molecularProfileIds = NA_character_,
  entrezGeneIds = NULL,
  sampleIds = NULL
)

mutationData(
  api,
  molecularProfileIds = NA_character_,
  entrezGeneIds = NULL,
  sampleIds = NULL
)

molecularData(
  api,
  molecularProfileIds = NA_character_,
  entrezGeneIds = NULL,
  sampleIds = NULL
)

searchOps(api, keyword)

samplesInSampleLists(api, sampleListIds = NA_character_)

sampleLists(api, studyId = NA_character_)

allSamples(api, studyId = NA_character_)

getSampleInfo(
  api,
  studyId = NA_character_,
  sampleListIds = NULL,
  projection = c("SUMMARY", "ID", "DETAILED", "META")
)

genePanels(api)

getGenePanel(api, genePanelId = NA_character_)

genePanelMolecular(
  api,
  molecularProfileId = NA_character_,
  sampleListId = NULL,
  sampleIds = NULL
)

getGenePanelMolecular(api, molecularProfileIds = NA_character_, sampleIds)

geneTable(api, pageSize = 1000, pageNumber = 0, ...)

queryGeneTable(
  api,
  by = c("entrezGeneId", "hugoGeneSymbol"),
  genes = NA_character_,
  genePanelId = NA_character_
)

getDataByGenes(
  api,
  studyId = NA_character_,
  genes = NA_character_,
  genePanelId = NA_character_,
  by = c("entrezGeneId", "hugoGeneSymbol"),
  molecularProfileIds = NULL,
  sampleListId = NULL,
  sampleIds = NULL,
  ...
)

Arguments

hostname

character(1) The internet location of the service (default: 'www.cbioportal.org')

protocol

character(1) The internet protocol used to access the hostname (default: 'https')

api.

character(1) The directory location of the API protocol within the hostname (default: '/api/api-docs')

token

character(1) The Authorization Bearer token e.g., "63eba81c-2591-4e15-9d1c-fb6e8e51e35d" or a path to text file.

api

An API object of class `cBioPortal` from the `cBioPortal` function

buildReport

logical(1) Indicates whether to append the build information to the `getStudies()` table (default FALSE)

studyId

character(1) Indicates the "studyId" as taken from `getStudies`

projection

character(default: "SUMMARY") Specify the projection type for data retrieval for details see API documentation

molecularProfileIds

character() A vector of molecular profile IDs

entrezGeneIds

numeric() A vector indicating entrez gene IDs

sampleIds

character() Sample identifiers

keyword

character(1) Keyword or pattern for searching through available operations

sampleListIds

character() A vector of 'sampleListId' as obtained from `sampleLists`

genePanelId

character(1) Identifies the gene panel, as obtained from the `genePanels` function

molecularProfileId

character(1) Indicates a molecular profile ID

sampleListId

character(1) A sample list identifier as obtained from `sampleLists()``

pageSize

numeric(1) The number of rows in the table to return

pageNumber

numeric(1) The pagination page number

...

Additional arguments to lower level API functions

by

character(1) Either 'entrezGeneId' or 'hugoGeneSymbol' for row metadata (default: 'entrezGeneId')

genes

character() Either Entrez gene identifiers or Hugo gene symbols. When included, the 'by' argument indicates the type of identifier provided and 'genePanelId' is ignored. Preference is given to Entrez IDs due to faster query responses.

Value

cBioPortal: An API object of class 'cBioPortal'

cBioPortalData: A data object of class 'MultiAssayExperiment'

API Metadata

* getStudies - Obtain a table of studies and associated metadata and optionally include a `buildReport` status (default FALSE) for each study. When enabled, the 'api_build' and 'pack_build' columns will be added to the table and will show if `MultiAssayExperiment` objects can be generated for that particular study identifier (`studyId`). The 'api_build' column corresponds to datasets obtained with `cBioPortalData` and the 'pack_build' column corresponds to datsets loaded via `cBioDataPack`.

* searchOps - Search through API operations with a keyword

* sampleLists - obtain all `sampleListIds` for a particular `studyId`

* allSamples - obtain all samples within a particular `studyId`

* genePanels - Show all available gene panels

* geneTable - Get a table of all genes by 'entrezGeneId' and 'hugoGeneSymbol'

* queryGeneTable - Get a table for only the `genes` or `genePanelId` of interest. Gene inputs are identified with the `by` argument

Patient Data

* clinicalData - Obtain clinical data for a particular study identifier ('studyId')

Molecular Profiles

* molecularProfiles - Produce a molecular profiles dataset for a given study identifier ('studyId')

Molecular Data

* fetchData - A convenience function to download both mutation and molecular data with `molecularProfileId`, `entrezGeneIds`, and `sampleIds`

* mutationData - Produce a dataset of mutation data using `molecularProfileId`, `entrezGeneIds`, and `sampleIds`

* molecularData - Produce a dataset of molecular profile data based on `molecularProfileId`, `entrezGeneIds`, and `sampleIds`

Sample Data

* samplesInSampleLists - get all samples associated with a 'sampleListId'

* getSampleInfo - Obtain sample metadata for a particular `studyId` or `sampleListId`

Gene Panels

* getGenePanels - Obtain the gene panel for a particular 'genePanelId'

* genePanelMolecular - get gene panel data for a particular `molecularProfileId` and either a vector of `sampleListId` or `sampleId`

* getGenePanelMolecular - get gene panel data for multiple `molecularProfileId`s and a vector of `sampleIds`

Genes

* getDataByGenes - Download data for a number of genes within `molecularProfileId` indicators, optionally a `sampleListId` can be provided.

Examples

cbio <- cBioPortal()

getStudies(api = cbio)
#> # A tibble: 365 × 13
#>    name    descr…¹ publi…² groups status impor…³ allSa…⁴ readP…⁵ studyId cance…⁶
#>    <chr>   <chr>   <lgl>   <chr>   <int> <chr>     <int> <lgl>   <chr>   <chr>  
#>  1 Adreno… "TCGA … TRUE    "PUBL…      0 2022-1…      92 TRUE    acc_tc… acc    
#>  2 Acute … "Compr… TRUE    "PUBL…      0 2022-1…      93 TRUE    all_st… bll    
#>  3 Hypodi… "Whole… TRUE    ""          0 2022-1…      44 TRUE    all_st… myeloid
#>  4 Adenoi… "Whole… TRUE    "ACYC…      0 2022-1…      12 TRUE    acbc_m… acbc   
#>  5 Adenoi… "Targe… TRUE    "ACYC…      0 2022-1…      28 TRUE    acyc_f… acyc   
#>  6 Adenoi… "Whole… TRUE    "ACYC…      0 2022-1…      25 TRUE    acyc_j… acyc   
#>  7 Adenoi… "WGS o… TRUE    "ACYC…      0 2022-1…     102 TRUE    acyc_m… acyc   
#>  8 Adenoi… "Whole… TRUE    "ACYC"      0 2022-1…      10 TRUE    acyc_m… acyc   
#>  9 Adenoi… "Whole… TRUE    "ACYC…      0 2022-1…      24 TRUE    acyc_s… acyc   
#> 10 Acute … "Whole… TRUE    "PUBL…      0 2022-1…      73 TRUE    all_st… bll    
#> # … with 355 more rows, 3 more variables: referenceGenome <chr>, pmid <chr>,
#> #   citation <chr>, and abbreviated variable names ¹​description, ²​publicStudy,
#> #   ³​importDate, ⁴​allSampleCount, ⁵​readPermission, ⁶​cancerTypeId

searchOps(api = cbio, keyword = "molecular")
#>  [1] "fetchGenePanelDataInMultipleMolecularProfilesUsingPOST"   
#>  [2] "getGenericAssayDataInMolecularProfileUsingGET"            
#>  [3] "fetchGenericAssayDataInMultipleMolecularProfilesUsingPOST"
#>  [4] "fetchGenericAssayDataInMolecularProfileUsingPOST"         
#>  [5] "fetchMolecularDataInMultipleMolecularProfilesUsingPOST"   
#>  [6] "getAllMolecularProfilesUsingGET"                          
#>  [7] "fetchMolecularProfilesUsingPOST"                          
#>  [8] "getMolecularProfileUsingGET"                              
#>  [9] "getDiscreteCopyNumbersInMolecularProfileUsingGET"         
#> [10] "fetchDiscreteCopyNumbersInMolecularProfileUsingPOST"      
#> [11] "getAllMolecularDataInMolecularProfileUsingGET"            
#> [12] "fetchAllMolecularDataInMolecularProfileUsingPOST"         
#> [13] "getMutationsInMolecularProfileBySampleListIdUsingGET"     
#> [14] "fetchMutationsInMolecularProfileUsingPOST"                
#> [15] "fetchMutationsInMultipleMolecularProfilesUsingPOST"       
#> [16] "getAllMolecularProfilesInStudyUsingGET"                   

## obtain clinical data
acc_clin <- clinicalData(api = cbio, studyId = "acc_tcga")
acc_clin
#> # A tibble: 92 × 85
#>    patie…¹ AGE   AJCC_…² ATYPI…³ CAPSU…⁴ CLIN_…⁵ CT_SC…⁶ CYTOP…⁷ DAYS_…⁸ DFS_M…⁹
#>    <chr>   <chr> <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 TCGA-O… 58    Stage … Atypic… Invasi… M0      [Unkno… Cytopl… 0       24.77  
#>  2 TCGA-O… 44    Stage … Atypic… Invasi… M1      YES     Cytopl… 0       9.49   
#>  3 TCGA-O… 23    Stage … Atypic… Invasi… M0      YES     Cytopl… 0       37.75  
#>  4 TCGA-O… 23    Stage … Atypic… Invasi… M1      YES     Cytopl… 0       4.14   
#>  5 TCGA-O… 30    Stage … Atypic… Invasi… M0      YES     Cytopl… 0       1.64   
#>  6 TCGA-O… 29    Stage … Atypic… Invasi… M0      YES     Cytopl… 0       88.80  
#>  7 TCGA-O… 30    Stage … Atypic… Invasi… M0      YES     Cytopl… 0       5.32   
#>  8 TCGA-O… 66    Stage … Atypic… Invasi… M0      YES     Cytopl… 0       17.41  
#>  9 TCGA-O… 22    Stage … Atypic… Invasi… M0      YES     Cytopl… 0       13.60  
#> 10 TCGA-O… 53    Stage … Atypic… Invasi… M1      YES     Cytopl… 0       NA     
#> # … with 82 more rows, 75 more variables: DFS_STATUS <chr>,
#> #   DIFFUSE_ARCHITECTURE <chr>, ETHNICITY <chr>, FORM_COMPLETION_DATE <chr>,
#> #   HISTOLOGICAL_DIAGNOSIS <chr>, HISTORY_ADRENAL_HORMONE_EXCESS <chr>,
#> #   HISTORY_BASIS_ADRENAL_HORMONE_DX <chr>, HISTORY_NEOADJUVANT_TRTYN <chr>,
#> #   HISTORY_OTHER_MALIGNANCY <chr>, ICD_10 <chr>, ICD_O_3_HISTOLOGY <chr>,
#> #   ICD_O_3_SITE <chr>, INFORMED_CONSENT_VERIFIED <chr>,
#> #   INITIAL_PATHOLOGIC_DX_YEAR <chr>, LATERALITY <chr>, …

molecularProfiles(api = cbio, studyId = "acc_tcga")
#> # A tibble: 9 × 8
#>   molecularAlterationType datatype name  descr…¹ showP…² patie…³ molec…⁴ studyId
#>   <chr>                   <chr>    <chr> <chr>   <lgl>   <lgl>   <chr>   <chr>  
#> 1 PROTEIN_LEVEL           LOG2-VA… Prot… Protei… FALSE   FALSE   acc_tc… acc_tc…
#> 2 PROTEIN_LEVEL           Z-SCORE  Prot… Protei… TRUE    FALSE   acc_tc… acc_tc…
#> 3 COPY_NUMBER_ALTERATION  DISCRETE Puta… Putati… TRUE    FALSE   acc_tc… acc_tc…
#> 4 COPY_NUMBER_ALTERATION  CONTINU… Capp… Capped… FALSE   FALSE   acc_tc… acc_tc…
#> 5 MUTATION_EXTENDED       MAF      Muta… Mutati… TRUE    FALSE   acc_tc… acc_tc…
#> 6 METHYLATION             CONTINU… Meth… Methyl… FALSE   FALSE   acc_tc… acc_tc…
#> 7 MRNA_EXPRESSION         CONTINU… mRNA… mRNA g… FALSE   FALSE   acc_tc… acc_tc…
#> 8 MRNA_EXPRESSION         Z-SCORE  mRNA… mRNA e… TRUE    FALSE   acc_tc… acc_tc…
#> 9 MRNA_EXPRESSION         Z-SCORE  mRNA… Log-tr… TRUE    FALSE   acc_tc… acc_tc…
#> # … with abbreviated variable names ¹​description, ²​showProfileInAnalysisTab,
#> #   ³​patientLevel, ⁴​molecularProfileId

genePanels(cbio)
#> # A tibble: 55 × 2
#>    description                                                           geneP…¹
#>    <chr>                                                                 <chr>  
#>  1 Targeted (27 cancer genes) sequencing of adenoid cystic carcinomas o… ACYC_F…
#>  2 Targeted panel of 232 genes.                                          Agilent
#>  3 Targeted panel of 8 genes.                                            AmpliS…
#>  4 ARCHER-HEME gene panel (199 genes)                                    ARCHER…
#>  5 ARCHER-SOLID Gene Panel (62 genes)                                    ARCHER…
#>  6 Targeted sequencing of various tumor types via bait v3.               bait_v3
#>  7 Targeted sequencing of various tumor types via bait v4.               bait_v4
#>  8 Targeted sequencing of various tumor types via bait v5.               bait_v5
#>  9 Targeted panel of 387 cancer-related genes.                           bcc_un…
#> 10 Research (CMO) IMPACT-Heme gene panel version 3.                      HemePA…
#> # … with 45 more rows, and abbreviated variable name ¹​genePanelId

(gp <- getGenePanel(cbio, "AmpliSeq"))
#> # A tibble: 8 × 2
#>   entrezGeneId hugoGeneSymbol
#>          <int> <chr>         
#> 1       171023 ASXL1         
#> 2         1788 DNMT3A        
#> 3         3417 IDH1          
#> 4         3418 IDH2          
#> 5        23451 SF3B1         
#> 6         6427 SRSF2         
#> 7         7157 TP53          
#> 8         7307 U2AF1         

muts <- mutationData(
    api = cbio,
    molecularProfileIds = "acc_tcga_mutations",
    entrezGeneIds = 1:1000,
    sampleIds = c("TCGA-OR-A5J1-01", "TCGA-OR-A5J2-01")
)
exps <- molecularData(
    api = cbio,
    molecularProfileIds = c("acc_tcga_rna_seq_v2_mrna", "acc_tcga_rppa"),
    entrezGeneIds = 1:1000,
    sampleIds = c("TCGA-OR-A5J1-01", "TCGA-OR-A5J2-01")
)

sampleLists(api = cbio, studyId = "acc_tcga")
#> # A tibble: 9 × 5
#>   category                                      name     descr…¹ sampl…² studyId
#>   <chr>                                         <chr>    <chr>   <chr>   <chr>  
#> 1 all_cases_with_mrna_rnaseq_data               Samples… Sample… acc_tc… acc_tc…
#> 2 all_cases_in_study                            All sam… All sa… acc_tc… acc_tc…
#> 3 all_cases_with_cna_data                       Samples… Sample… acc_tc… acc_tc…
#> 4 all_cases_with_mutation_and_cna_data          Samples… Sample… acc_tc… acc_tc…
#> 5 all_cases_with_mutation_and_cna_and_mrna_data Complet… Sample… acc_tc… acc_tc…
#> 6 all_cases_with_methylation_data               Samples… Sample… acc_tc… acc_tc…
#> 7 all_cases_with_methylation_data               Samples… Sample… acc_tc… acc_tc…
#> 8 all_cases_with_rppa_data                      Samples… Sample… acc_tc… acc_tc…
#> 9 all_cases_with_mutation_data                  Samples… Sample… acc_tc… acc_tc…
#> # … with abbreviated variable names ¹​description, ²​sampleListId

samplesInSampleLists(
    api = cbio,
    sampleListIds = c("acc_tcga_rppa", "acc_tcga_cnaseq")
)
#> CharacterList of length 2
#> [["acc_tcga_cnaseq"]] TCGA-OR-A5J1-01 TCGA-OR-A5J2-01 ... TCGA-PK-A5HC-01
#> [["acc_tcga_rppa"]] TCGA-OR-A5J2-01 TCGA-OR-A5J3-01 ... TCGA-PK-A5HA-01

genePanels(api = cbio)
#> # A tibble: 55 × 2
#>    description                                                           geneP…¹
#>    <chr>                                                                 <chr>  
#>  1 Targeted (27 cancer genes) sequencing of adenoid cystic carcinomas o… ACYC_F…
#>  2 Targeted panel of 232 genes.                                          Agilent
#>  3 Targeted panel of 8 genes.                                            AmpliS…
#>  4 ARCHER-HEME gene panel (199 genes)                                    ARCHER…
#>  5 ARCHER-SOLID Gene Panel (62 genes)                                    ARCHER…
#>  6 Targeted sequencing of various tumor types via bait v3.               bait_v3
#>  7 Targeted sequencing of various tumor types via bait v4.               bait_v4
#>  8 Targeted sequencing of various tumor types via bait v5.               bait_v5
#>  9 Targeted panel of 387 cancer-related genes.                           bcc_un…
#> 10 Research (CMO) IMPACT-Heme gene panel version 3.                      HemePA…
#> # … with 45 more rows, and abbreviated variable name ¹​genePanelId

getGenePanel(api = cbio, genePanelId = "IMPACT341")
#> # A tibble: 341 × 2
#>    entrezGeneId hugoGeneSymbol
#>           <int> <chr>         
#>  1           25 ABL1          
#>  2        84142 ABRAXAS1      
#>  3          207 AKT1          
#>  4          208 AKT2          
#>  5        10000 AKT3          
#>  6          238 ALK           
#>  7          242 ALOX12B       
#>  8       139285 AMER1         
#>  9          324 APC           
#> 10          367 AR            
#> # … with 331 more rows

queryGeneTable(api = cbio, by = "entrezGeneId", genes = 7157)
#> # A tibble: 1 × 3
#>   entrezGeneId hugoGeneSymbol type          
#>          <int> <chr>          <chr>         
#> 1         7157 TP53           protein-coding

getDataByGenes(
    cbio, studyId = "acc_tcga", genes = 1:3,
    by = c("entrezGeneId", "hugoGeneSymbol"),
    molecularProfileId = "acc_tcga_rppa",
    sampleListId = "acc_tcga_rppa"
)
#> named list()