A class for representing the cBioPortal API protocol

The cBioPortal class is a representation of the cBioPortal API protocol that directly inherits from the Service class in the AnVIL package. For more information, see the AnVIL package.

# S4 method for cBioPortal
operations(x, ..., .deprecated = FALSE)

Arguments

x

A Service instance or API representation as given by the cBioPortal function.

...

additional arguments passed to methods or, for `operations,Service-method`, to the internal `get_operation()` function.

.deprecated

optional logical(1) include deprecated operations?

Details

This class takes the static API as provided at https://www.cbioportal.org/api/api-docs and creates an R object with the help from underlying infrastructure (i.e., rapiclient and AnVIL) to give the user a unified representation of the API specification provided by the cBioPortal group. Users are not expected to interact with this class other than to use it as input to the functionality provided by the rest of the package.

Functions

operations,cBioPortal-method:

See also

Examples

cBioPortal()
#> service: cBioPortal
#> tags(); use cbioportal$<tab completion>:
#> # A tibble: 61 × 3
#>    tag                 operation                                  summary       
#>    <chr>               <chr>                                      <chr>         
#>  1 Cancer Types        getAllCancerTypesUsingGET                  Get all cance…
#>  2 Cancer Types        getCancerTypeUsingGET                      Get a cancer …
#>  3 Clinical Attributes fetchClinicalAttributesUsingPOST           Fetch clinica…
#>  4 Clinical Attributes getAllClinicalAttributesInStudyUsingGET    Get all clini…
#>  5 Clinical Attributes getAllClinicalAttributesUsingGET           Get all clini…
#>  6 Clinical Attributes getClinicalAttributeInStudyUsingGET        Get specified…
#>  7 Clinical Data       fetchAllClinicalDataInStudyUsingPOST       Fetch clinica…
#>  8 Clinical Data       fetchClinicalDataUsingPOST                 Fetch clinica…
#>  9 Clinical Data       getAllClinicalDataInStudyUsingGET          Get all clini…
#> 10 Clinical Data       getAllClinicalDataOfPatientInStudyUsingGET Get all clini…
#> # … with 51 more rows
#> tag values:
#>   Cancer Types, Clinical Attributes, Clinical Data, Copy Number
#>   Segments, Discrete Copy Number Alterations, Gene Panels, Generic
#>   Assays, Genes, Molecular Data, Molecular Profiles, Mutations,
#>   Patients, Sample Lists, Samples, Structural Variants, Studies,
#>   Treatments
#> schemas():
#>   AlleleSpecificCopyNumber, AlterationFilter,
#>   AndedPatientTreatmentFilters, AndedSampleTreatmentFilters,
#>   CancerStudy
#>   # ... with 58 more elements