Download data from the cBioPortal API

Obtain a MultiAssayExperiment object for a particular gene panel, studyId, molecularProfileIds, and sampleListIds combination. Default molecularProfileIds and sampleListIds are set to NULL for including all data. This option is best for users who wish to obtain a section of the study data that pertains to a specific molecular profile and gene panel combination. For users looking to download the entire study data as provided by the https://cbioportal.org/datasets, refer to cBioDataPack.

cBioPortalData(
  api,
  studyId = NA_character_,
  genePanelId = NA_character_,
  genes = NA_character_,
  molecularProfileIds = NULL,
  sampleListId = NULL,
  sampleIds = NULL,
  by = c("entrezGeneId", "hugoGeneSymbol")
)

Arguments

api
An API object of class `cBioPortal` from the `cBioPortal` function

studyId
character(1) Indicates the "studyId" as taken from `getStudies`

genePanelId
character(1) Identifies the gene panel, as obtained from the `genePanels` function

genes
character() Either Entrez gene identifiers or Hugo gene symbols. When included, the 'by' argument indicates the type of identifier provided and 'genePanelId' is ignored. Preference is given to Entrez IDs due to faster query responses.

molecularProfileIds
character() A vector of molecular profile IDs

sampleListId
character(1) A sample list identifier as obtained from `sampleLists()``

sampleIds
character() Sample identifiers

by
character(1) Either 'entrezGeneId' or 'hugoGeneSymbol' for row metadata (default: 'entrezGeneId')

Value

A MultiAssayExperiment object

Details

As of May 2020, there were about 96.6 percent of the 268 datasets successfully imported. The datasets that currently fail to import are:

        c("all_stjude_2015", "sclc_ucologne_2015", "skcm_ucla_201
        "sclc_jhu", "gbm_tcga_pub2013", "hnsc_tcga_pub", "kirc_tc
        "brca_tcga_pub", "brca_tcga_pub2015")
    

Note that changes to the cBioPortal API may affect this rate at any time. If you encounter any issues, please open a GitHub issue at the https://github.com/waldronlab/cBioPortalData/issues/ page with a fully reproducible example.

See also

Examples

cbio <- cBioPortal()

samps <- samplesInSampleLists(cbio, "acc_tcga_rppa")[[1]]

getGenePanelMolecular(
    cbio, molecularProfileIds = c("acc_tcga_rppa", "acc_tcga_linear_CNA"),
    samps
)

#> # A tibble: 92 x 7
#>    uniqueSampleKey  uniquePatientKey molecularProfil… sampleId patientId studyId
#>    <chr>            <chr>            <chr>            <chr>    <chr>     <chr>  
#>  1 VENHQS1PUi1BNUo… VENHQS1PUi1BNUo… acc_tcga_linear… TCGA-OR… TCGA-OR-… acc_tc…
#>  2 VENHQS1PUi1BNUo… VENHQS1PUi1BNUo… acc_tcga_linear… TCGA-OR… TCGA-OR-… acc_tc…
#>  3 VENHQS1PUi1BNUo… VENHQS1PUi1BNUo… acc_tcga_linear… TCGA-OR… TCGA-OR-… acc_tc…
#>  4 VENHQS1PUi1BNUo… VENHQS1PUi1BNUo… acc_tcga_linear… TCGA-OR… TCGA-OR-… acc_tc…
#>  5 VENHQS1PUi1BNUo… VENHQS1PUi1BNUo… acc_tcga_linear… TCGA-OR… TCGA-OR-… acc_tc…
#>  6 VENHQS1PUi1BNUo… VENHQS1PUi1BNUo… acc_tcga_linear… TCGA-OR… TCGA-OR-… acc_tc…
#>  7 VENHQS1PUi1BNUp… VENHQS1PUi1BNUp… acc_tcga_linear… TCGA-OR… TCGA-OR-… acc_tc…
#>  8 VENHQS1PUi1BNUp… VENHQS1PUi1BNUp… acc_tcga_linear… TCGA-OR… TCGA-OR-… acc_tc…
#>  9 VENHQS1PUi1BNUp… VENHQS1PUi1BNUp… acc_tcga_linear… TCGA-OR… TCGA-OR-… acc_tc…
#> 10 VENHQS1PUi1BNUp… VENHQS1PUi1BNUp… acc_tcga_linear… TCGA-OR… TCGA-OR-… acc_tc…
#> # … with 82 more rows, and 1 more variable: profiled <lgl>

acc_tcga <- cBioPortalData(
    cbio, by = "hugoGeneSymbol",
    studyId = "acc_tcga",
    genePanelId = "AmpliSeq",
    molecularProfileIds =
        c("acc_tcga_rppa", "acc_tcga_linear_CNA", "acc_tcga_mutations")
)

#> Error in eval(args[["api"]]): object 'cbio' not found